+++ store/store_test.go
@@ -128,6 +128,21 @@ func TestIteratorCommitAndPrefixed(t *testing.T) {
 	it4.Close()
 }
 
+func TestUnsafeRollback(t *testing.T) {
+	store, _, cleanup := testStore(t)
+	defer cleanup()
+	for i := 0; i < 100; i++ {
+		require.NoError(t, store.Set([]byte("key"), []byte(fmt.Sprintf("%d", i+1))))
+		_, err := store.Commit()
+		require.NoError(t, err)
+	}
+	require.Equal(t, uint64(100), store.Version())
+	store.UnsafeRollback(1)
+	k, err := store.Get([]byte("key"))
+	require.NoError(t, err)
+	require.Equal(t, "1", string(k))
+}
+
 //func TestProof(t *testing.T) {
 //	store, _, cleanup := testStore(t)
 //	defer cleanup()